flask_app/requirements.txt
Flask
numpy
pillow
tensorflow